# Conflicts:
#	src/main/java/ru/netology/services/StatCalcService.java

+++ src/test/java/ru/netology/services/StatCalcServiceTest.java
@@ -26,8 +26,8 @@ public class StatCalcServiceTest {
             8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18
         };
 
-        long expected = 15;
-        long actual = calculator.calcAverage(inputArray);
+        int expected = 15;
+        int actual = calculator.calcAverage(inputArray);
 
         Assertions.assertEquals(expected, actual);
     }
